better organization and naming for new syntax highlighting support
