+++ lib/postmarkdown.rb
@@ -3,3 +3,4 @@ require 'postmarkdown/config'
 require 'postmarkdown/routes'
 require 'postmarkdown/railtie'
 require 'postmarkdown/util'
+require 'postmarkdown/markdown_renderer'
